feat(auth): global/jwt jwt 폴더 생성
